Refactor check_only_instances() functionality
The purpose of the `check_only_instances()` function is to ensure that
only one instance of the script runs at a time, and the purpose of this
change is to increase portability of the 'backup' script.

The `pidof` command on Debian has different options than the BSD variant.
The flock(1) functionality can achieve the same result with increased
portability.

The `flock` call at the top of the script uses the 'backup' script
itself as a file lock, fulfilling the single instance criteria by
disallowing other instances of the script to try and claim that same
lock.

For more information, see the man page for flock(1).